Enable sketching on the grid via mouse
Each square the mouse touches turns from red to black. Sadly, this works
in preview mode but does not work on my laptop, I assume because it is
a mouse + touchscreen affair. On the laptop, I have to click the mouse
to get the square to turn black, which makes it more Lite Brite and less
Etch-a-Sketch.