Update to gradient-based sketching
As the user moves their mouse over the canvas, each square the pointer
touches becomes a little darker. The squares start at a 10% gray and get
darker in even 10% increments up to black at 100%. Hovering on a square
does not cause ongoing darkening. The pointer has to leave the square
and return to it to add the next 10%. This results in a relatively
steady and controllable sketching experience.